backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Portafolio Profesional"
authors = ["Pablo <godoypablojuan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
